@@ -1,7 +1,6 @@
 package com.ascargon.rocketshow.audio;
 
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.annotation.XmlRootElement;
 
 /**
  * An audio bus containing a number of channels.
@@ -11,24 +10,24 @@ import javax.xml.bind.annotation.XmlRootElement;
 @XmlRootElement
 public class AudioBus {
 
-	private String name;
+    private String name;
 
-	private int channels = 2;
+    private int channels = 2;
 
-	public String getName() {
-		return name;
-	}
+    public String getName() {
+        return name;
+    }
 
-	public void setName(String name) {
-		this.name = name;
-	}
+    public void setName(String name) {
+        this.name = name;
+    }
 
-	public int getChannels() {
-		return channels;
-	}
+    public int getChannels() {
+        return channels;
+    }
 
-	public void setChannels(int channels) {
-		this.channels = channels;
-	}
+    public void setChannels(int channels) {
+        this.channels = channels;
+    }
 
 }
