@@ -2,7 +2,7 @@ package com.ascargon.rocketshow.audio;
 
 import com.ascargon.rocketshow.composition.CompositionFile;
 
-import javax.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlElement;
 
 public class AudioCompositionFile extends CompositionFile {
 
