@@ -1,16 +1,6 @@
 package com.ascargon.rocketshow.composition;
 
-import com.ascargon.rocketshow.audio.AudioCompositionFile;
-import com.ascargon.rocketshow.midi.MidiCompositionFile;
-import com.ascargon.rocketshow.video.VideoCompositionFile;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlElementWrapper;
-import javax.xml.bind.annotation.XmlElements;
-import javax.xml.bind.annotation.XmlRootElement;
-import java.util.ArrayList;
-import java.util.List;
+import jakarta.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement
 public class LeadSheet {
