@@ -3,8 +3,8 @@ package com.ascargon.rocketshow.midi;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlElementWrapper;
+import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlElementWrapper;
 
 import com.ascargon.rocketshow.composition.CompositionFile;
 import com.fasterxml.jackson.annotation.JsonTypeName;
