@@ -1,7 +1,7 @@
 package com.ascargon.rocketshow.raspberry;
 
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlRootElement;
 
 import com.ascargon.rocketshow.util.ControlAction;
 
