@@ -1,7 +1,6 @@
 package com.ascargon.rocketshow.util;
 
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.annotation.XmlRootElement;
 
 /**
  * Container for the used and available disc space.
@@ -11,21 +10,23 @@ import javax.xml.bind.annotation.XmlRootElement;
 @XmlRootElement
 public class DiskSpace {
 
-	private double usedMB = 0;
-	private double availableMB = 0;
+    private double usedMB = 0;
+    private double availableMB = 0;
 
-	public double getUsedMB() {
-		return usedMB;
-	}
-	public void setUsedMB(double usedMB) {
-		this.usedMB = usedMB;
-	}
+    public double getUsedMB() {
+        return usedMB;
+    }
 
-	public double getAvailableMB() {
-		return availableMB;
-	}
-	public void setAvailableMB(double availableMB) {
-		this.availableMB = availableMB;
-	}
+    public void setUsedMB(double usedMB) {
+        this.usedMB = usedMB;
+    }
+
+    public double getAvailableMB() {
+        return availableMB;
+    }
+
+    public void setAvailableMB(double availableMB) {
+        this.availableMB = availableMB;
+    }
 
 }
