@@ -323,7 +323,7 @@ public class DefaultDesignerService implements DesignerService {
             // Mix the existing value with the new value
             newValue = existingValue * (1 - intensityPercentage) + newValue * intensityPercentage;
 
-            logger.info("existingValue: " + existingValue + ", intensityPercentage: " + intensityPercentage + ", newValue: " + newValue);
+            logger.trace("existingValue: " + existingValue + ", intensityPercentage: " + intensityPercentage + ", newValue: " + newValue);
         }
 
         // Remove the existent value, if available
@@ -1108,6 +1108,7 @@ public class DefaultDesignerService implements DesignerService {
     public void load(CompositionPlayer compositionPlayer, Project project, Pipeline pipeline) {
         this.compositionPlayer = compositionPlayer;
         this.project = project;
+        this.pipeline = pipeline;
 
         if (compositionPlayer != null) {
             this.composition = getCompositionByName(project, compositionPlayer.getComposition().getName());
